feat: 상품 조회 기능, 테스트 구현
